# gs: add option to set chunk size