@fd0 2017-10-16
Oh, good point, thanks for digging into the problem!

I'd like to avoid exposing this option to end users, is there any harm in using 256KiB by default?

From the backend, it's not possible to directly see if the upload limit is configured, so we'd need to pass that in when opening the backend. The upload limit divided by four or eight should be a good heuristic.

What do you think?