@prattmic 2017-10-17
Actually, looking into it more, I think it might be reasonable to set the chunk size to 0 by default.

The GCS library is doing this chunking to implement [resumable uploads](https://cloud.google.com/storage/docs/json_api/v1/how-tos/resumable-upload). It uploads the object in chunk size chunks and has automatic retry and exponential backoff if chunk upload fails.

The main purpose of resumable uploads is to avoid having to re-upload lots of data if a request fails. (Max re-upload is the chunk size). However, the default chunk size is 8MB, and if I understand correctly, the maximum pack size is 12MB (4MB min pack size + 8MB max blob size), so the default chunk size isn't much more than the maximum object size we'll upload anyways.

Setting the chunk size to 0 disables resumable uploads altogether and uploads the entire object in a single request. The passed Reader is used more directly, without explicit large buffers.

This wouldn't eliminate all buffering; the GCS libraries still io.Copy the Reader to an io.Pipe(), and net/http does a fair amount of io.Copy'ing itself, but I suspect this would get the best rate limiting behavior we can easily achieve.

Thoughts?