@fd0 2017-10-17
I think we can set the chunk size to zero.